Catch up (up to chapter 4) and some refactoring; add bullet points for chapter 3

+++ src/com/josh/headfirst/chapter3/Triangle.java
@@ -8,7 +8,7 @@ class Triangle
 
     void setArea()
     {
-        area = (height * length) / 2;
+        area = (double) (height * length) / 2;
     }
 
     /* MAIN METHOD */
